Api Documentation

+++ app/controllers/v1/yachts_controller.rb
@@ -1,45 +1,44 @@
 class V1::YachtsController < ApplicationController
   before_action :set_yacht, only: %i[show update destroy]
-  # before_action :authenticate_user!
+  before_action :authenticate_user!
 
   # GET /yachts
   def index
     @yachts = Yacht.all
 
-    render json: @yachts
+    render json: YachtSerializer.new(@yachts).serializable_hash[:data], status: :ok
   end
 
   # GET /yachts/1
   def show
-    render json: YachtSerializer.new(@yacht).serializable_hash[:data][:attributes]
+    render json: YachtSerializer.new(@yacht).serializable_hash[:data][:attributes], status: :ok
   end
 
   # POST /yachts
   def create
-    @yacht = Yacht.new(yacht_params)
+    if current_user.admin?
+      @yacht = Yacht.new(yacht_params)
 
-    if @yacht.save
-      render json: YachtSerializer.new(@yacht).serializable_hash[:data][:attributes], status: :created
+      if @yacht.save
+        render json: YachtSerializer.new(@yacht).serializable_hash[:data][:attributes], status: :created
+      else
+        render json: @yacht.errors, status: :unprocessable_entity
+      end
     else
-      render json: @yacht.errors, status: :unprocessable_entity
+      render json: { message: 'You are not authorized to perform this action' }, status: :unauthorized
     end
   end
 
-  # PATCH/PUT /yachts/1
-  # def update
-  #   if @yacht.update(yacht_params)
-  #     render json: @yacht
-  #   else
-  #     render json: @yacht.errors, status: :unprocessable_entity
-  #   end
-  # end
-
   # DELETE /yachts/1
   def destroy
-    if @yacht.destroy
-      render json: { message: 'Yacht deleted' }, status: 200
+    if current_user.admin?
+      if @yacht.destroy
+        render json: { message: 'Yacht deleted' }, status: 200
+      else
+        render json: { message: 'Yacht could not be deleted' }, status: 500
+      end
     else
-      render json: { message: 'Yacht could not be deleted' }, status: 500
+      render json: { message: 'You are not authorized to perform this action' }, status: :unauthorized
     end
   end
 
@@ -47,7 +46,8 @@ class V1::YachtsController < ApplicationController
 
   # Use callbacks to share common setup or constraints between actions.
   def set_yacht
-    @yacht = Yacht.find(params[:id])
+    @yacht = Yacht.find_by_id(params[:id])
+    render json: { message: 'Yacht not found' }, status: 404 if @yacht.nil?
   end
 
   # Only allow a list of trusted parameters through.
